backend/requirements.txt
fastapi==0.109.0
uvicorn[standard]==0.27.0
tortoise-orm==0.20.0
passlib[bcrypt]==1.7.4
bcrypt==4.2.1
PyJWT[crypto]>=2.8.0
python-multipart==0.0.6
asyncpg==0.29.0
pandas==2.1.4
openpyxl==3.1.2
httpx==0.27.0
